ceph: start the cluster monitoring goroutines immediately after restart
The monitoring goroutines that check for mon and osd health,
the ceph cluster status, and the bucket provisioner should
be started immediately with the first reconcile after an
operator restart instead of waiting until the first
reconcile is completed. This way the cluster health will
be monitored even during the first reconcile.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>