ceph: helm chart fix for rook-ceph
Go's template language doesn't treat infix `and` as a short-circuit,
which led to fatal errors after an update in Go 1.14 surfaced that as
a fatal error.  This commit removes the infix `and` and replaces it with
a default dict to handle the case of a missing Heml value.

Closes: https://github.com/rook/rook/issues/5378
Signed-off-by: Judson <nyarly@gmail.com>
(cherry picked from commit f89ecb615695ffee602b76156b7b94104ee78dff)