ceph: save mon config even if failed to remove dead mon
During mon failover if the mon fails to be removed, the new mon
config was not being saved. The mon removal anyway is not retried
if it fails, so it should make a best effort to remove all the
resources related to the mon, and then always update the mon
config after the failover.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 0be6fdfa2d28a78347075c06975911f74e9a87fe)