ceph: during mon failover remove the backing pvc
If the mon is backed by a pvc, ensure the pvc is deleted when
the mon is failed over. The pvc is being left in a bound state
even though it will no longer be used in the cluster.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 1d370b1314a1b9c28ec90b788b136c03bd9bdc40)