ceph: force delete stuck mons in every reconcile
The stuck mons were only being force deleted in the mon health checker.
After an operator restart, the health check may not be executed
for 20 minutes until after the timeout during the main reconcile
which is holding onto a lock for mon reconcile. Now the stuck mon
will be deleted during the reconcile as well.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>