ceph: during upgrade service selector cannot be changed
The service selector labels cannot be updated during the upgrade or else the
mon commands will not respond until the mon deployment is also updated.
Since there is a mon command to check for the ceph versions before the
deployment is updated, the operator would remain stuck during the upgrade.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>