ci: update K8s version for stability
K8s v1.16.13 is used in CI. In this version, there is a serious bug
which causes a continuous restart of both kube-scheduler
and kube-controller-manager. It results in the instability of CI.

This bug is fixed in v1.16.14 or later by the following PR.
https://github.com/kubernetes/kubernetes/pull/93208

Both 6415 and 6390, which only happen in v1.16.13, would be
caused by this issue.

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>

docs: ceph disaster recovery

In this version, the paranthesis has been removed around the arguments for the ceph-mon command.

Note on removing the paranthesis additional to log-std-err has been added.