docs: precise three nodes
When following the quick start, it may not work, because the user don't know that he need three nodes. And three nodes may also not be correct, because the control plane is also a node. So we should precise what the user need to run his cluster correctly.

Closes: https://github.com/rook/rook/issues/6540
Signed-off-by: Gerhard Seidel <gseidel.message@googlemail.com>
(cherry picked from commit 042e10bcf6ab3bb69506ce973549cecdb1057cc3)