docs: remove inactive member from steering committee
Dmitry is no longer participating in Rook and has moved on
from the EdgeFS team. The position on the steering committee
is now open. Per the governance, we are looking for an owner
of a storage provider other than Ceph who has overall
perspective and interest in taking Rook forward.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>